Handle low-data GTI
Occasionally, due to data delivery timing, the most recent GTI will only continue a single data point, which messes up the trend fitting code and throws an error. This is a quick fix that will ignore the GTI if it contains less than two data points (the most recent GTI will be revisited in a subsequent clockfile run anyway, so should not cause any problems).